+++ lighthouse-core/test/lib/asset-saver-test.js
@@ -14,7 +14,7 @@ import traceEvents from '../fixtures/traces/progressive-app.json';
 import dbwTrace from '../results/artifacts/defaultPass.trace.json';
 import dbwResults from '../results/sample_v2.json';
 import Audit from '../../audits/audit.js';
-import {getModuleDirectory} from '../../scripts/esm-utils.js';
+import {getModuleDirectory} from '../../../esm-utils.mjs';
 import fullTraceObj from '../fixtures/traces/progressive-app-m60.json';
 import devtoolsLog from '../fixtures/traces/progressive-app-m60.devtools.log.json';
