+++ lighthouse-core/test/lib/i18n/i18n-test.js
@@ -10,7 +10,7 @@ import {jest} from '@jest/globals';
 import log from 'lighthouse-logger';
 
 import i18n from '../../../lib/i18n/i18n.js';
-import {getModuleDirectory} from '../../../scripts/esm-utils.js';
+import {getModuleDirectory} from '../../../../esm-utils.mjs';
 
 const moduleDirectory = getModuleDirectory(import.meta);
